+++ contracts/tests/resolvers/TestASDataResolver.sol
@@ -12,7 +12,7 @@ contract TestASDataResolver is ASResolver {
         address, /* recipient */
         bytes calldata, /* schema */
         bytes calldata data,
-        uint256, /* expirationTime */
+        uint32, /* expirationTime */
         address /* msgSender */
     ) external payable virtual override returns (bool) {
         // Verifies that the data is either 0 or 1.
